review: move non-tests to conftest
